requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
PyYAML==6.0.2
requests==2.32.3
psutil==6.0.0
reportlab>=4.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
aiofiles>=23.0.0
